+++ DHLWebAPI/Models/TblAddress.cs
@@ -11,7 +11,7 @@ namespace DHLWebAPI.Models
         }
 
         public int IdAddress { get; set; }
-        public byte IdAddressType { get; set; }
+        public int IdAddressType { get; set; }
         public string AddressLabel { get; set; }
         public string Country { get; set; }
         public string Province { get; set; }
